em-backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pyjwt==2.9.0
python-multipart==0.0.9
elevenlabs==2.16.0
gspread==6.1.4
oauth2client==4.1.3
notion-client==2.2.1
python-dotenv==1.0.0
requests==2.31.0
